refactor: internal_domain_logicパッケージのバレルファイルを削除
- internal_domain_logic.dartバレルファイルを削除
- src/operational_settingsをlib/operational_settingsに移動
- src/theme_settingをlib/theme_settingに移動
- 不要なinternal_domain_logic_base.dartとinternal_domain_logic_test.dartを削除
- 依存ファイルのimport文を更新
