+++ apps/app/lib/composition_root/repositories/operational_settings_repository.dart
@@ -1,5 +1,5 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:internal_domain_logic/internal_domain_logic.dart';
+import 'package:internal_domain_logic/operational_settings/operational_settings_repository.dart';
 import 'package:internal_infrastructure/operational_settings/operational_settings_repository_impl.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
