--- packages/domain_logic/lib/internal_domain_logic.dart
@@ -1,10 +1,0 @@
-/// Support for doing something awesome.
-///
-/// More dartdocs go here.
-library;
-
-export 'src/internal_domain_logic_base.dart';
-export 'src/operational_settings/operational_settings_repository.dart';
-export 'src/theme_setting/theme_setting_repository.dart';
-
-// TODO: Export any libraries intended for clients of this package.
