+++ packages/infrastructure/lib/operational_settings/operational_settings_repository_impl.dart
@@ -1,4 +1,4 @@
-import 'package:internal_domain_logic/internal_domain_logic.dart';
+import 'package:internal_domain_logic/operational_settings/operational_settings_repository.dart';
 import 'package:internal_domain_model/internal_domain_model.dart';
 import 'package:pub_semver/pub_semver.dart';
