+++ packages/infrastructure/lib/theme_setting/theme_setting_repository_impl.dart
@@ -1,4 +1,4 @@
-import 'package:internal_domain_logic/internal_domain_logic.dart';
+import 'package:internal_domain_logic/theme_setting/theme_setting_repository.dart';
 import 'package:internal_domain_model/internal_domain_model.dart';
 import 'package:internal_infrastructure/shared_preferences/shared_preference_data_source.dart';
